NMS-12445: Move supported properties to api module
