+++ common/src/main/java/org/opennms/integration/api/v1/graph/immutables/ImmutableGraph.java
@@ -40,6 +40,7 @@ import java.util.stream.Collectors;
 import org.opennms.integration.api.v1.graph.Edge;
 import org.opennms.integration.api.v1.graph.Graph;
 import org.opennms.integration.api.v1.graph.GraphInfo;
+import org.opennms.integration.api.v1.graph.Properties;
 import org.opennms.integration.api.v1.graph.Vertex;
 import org.opennms.integration.api.v1.graph.VertexRef;
 import org.opennms.integration.api.v1.util.ImmutableCollections;
